modelo Comment, comment.controller, foreing Key terminado

+++ controllers/movie.controller.js
@@ -1,25 +1,24 @@
 const Movie = require('../models/Movie');
+const Comment = require('../models/Comment');
+
 
 
 //OBTENER TODAS LAS PELICULAS
-const getMovies = (req, res) => {
+const getTheMovies = (req, res) => {
 
   Movie.findAndCountAll({
 
-    attributes: ['id', 'title', 'year']
+    attributes: ['id', 'title', 'year', 'image', 'score']
 
   }).then( movies => {
 
-    res.json({
-      ok: true,
-      movies: movies
-    })
+    res.json(movies)
 
   }).catch( err => {
 
     res.json({
       ok: false,
-      msg: err
+      msg: 'Error inesperado...'
     });
 
   })
@@ -27,7 +26,7 @@ const getMovies = (req, res) => {
 }
 
 //CREAR UNA PELICULA
-const createMovie = (req, res) => {
+const createTheMovie = (req, res) => {
 
   const { title, year, image } = req.body;
 
@@ -76,11 +75,11 @@ const createMovie = (req, res) => {
 }
 
 //OBTENER UNA PELICULA
-const getMovie = (req, res) => {
+const getTheMovie = (req, res) => {
 
   Movie.findByPk(
     req.params.id, 
-    {attributes: ['id', 'title', 'year', 'image']}   
+    {attributes: ['id', 'title', 'year', 'image', 'score']}   
 
   ).then( movie => {
 
@@ -91,10 +90,7 @@ const getMovie = (req, res) => {
       })
 
     } else {
-      res.json({
-        ok: true,
-        movie
-      })
+      res.json(movie);
 
     }
 
@@ -102,7 +98,7 @@ const getMovie = (req, res) => {
 
     res.json({
       ok: false,
-      msg: err
+      msg: 'Error inesperado...'
     });
 
   })
@@ -110,7 +106,7 @@ const getMovie = (req, res) => {
 }
 
 //EDITAR UNA PELICULA
-const updateMovie = (req, res) => {
+const updateTheMovie = (req, res) => {
 
   Movie.update({
     title: req.body.title,
@@ -151,25 +147,100 @@ const updateMovie = (req, res) => {
 }
 
 //BORRAR UNA PELICULA
-const deleteMovie = (req, res) => {
+const deleteTheMovie = (req, res) => {
 
-  Movie.destroy({
+  //borrar comentarios de la pelicula
+  Comment.destroy({
+
     where: {
-      id: req.params.id
+      movieId: req.params.id
     }
 
-  }).then( result => {
+  //resp devuelve un numero que representa la cantidad de items borrados
+  }).then( resp => {
+    //si es 0 es que la pelicula no tenia comentarios
+    if (resp >= 0) {
 
-    if (result === 1) {
+      //borrar pelicula
+      Movie.destroy({
+        where: {
+          id: req.params.id
+        }
+
+      }).then( result => {
+
+        if (result >= 1) {
+          res.json({
+            ok: true,
+            msg: 'pelicula borrada con exito.'
+          })
+
+        } else {
+          res.json({
+            ok: false,
+            msg: 'no se pudo borrar la pelicula.'
+          })
+
+        }
+
+      }).catch( err => {
+
+        res.json({
+          ok: false,
+          msg: 'Error inesperado...'
+        });
+
+      })
+
+      
+    } else {
+
       res.json({
-        ok: true,
-        msg: 'pelicula borrada con exito.'
+        ok: false,
+        msg: 'Error inesperado... no se pudieron borrar los comentarios.'
+      });
+      
+    }
+
+  });
+
+
+}
+
+//OBTENER UNA PELICULA CON SUS COMENTARIOS
+const getMovieAndComments = (req, res) => {
+
+  Movie.findByPk(
+    req.params.id, 
+    {attributes: ['id', 'title', 'year', 'image', 'score']}   
+
+  ).then( movie => {
+
+    if (!movie) {
+      res.json({
+        ok: false,
+        msg: 'pelicula no encontrada.'
       })
 
     } else {
-      res.json({
-        ok: false,
-        msg: 'no se pudo borrar la pelicula.'
+  
+      movie.getComments({ attributes: ['comment'] }).then( comments => {
+
+        //creamos un array solo con los comentarios
+        const onlyComments = comments.map( obj => obj = obj.comment);
+        
+        //creamos un nuevo objeto con la pelicula y sus comentarios
+        const commentedMovie = {
+            id: movie.id,
+            title: movie.title,
+            year: movie.year,
+            image: movie.image,
+            score: movie.score,
+            comments: onlyComments
+        }
+        
+        res.json(commentedMovie);
+
       })
 
     }
@@ -178,7 +249,57 @@ const deleteMovie = (req, res) => {
 
     res.json({
       ok: false,
-      msg: err
+      msg: 'Error inesperado...'
+    });
+
+  })
+
+}
+
+//CALIFICAR Y COMENTAR PELICULA
+const rateAndCommentMovie = (req, res) => {
+
+  const { rate, comment } = req.body;
+  const movieId = req.params.id;
+
+  //Creamos el comentario con la llave foranea
+  Comment.create({
+    rate: rate,
+    comment: comment,
+    movieId: movieId
+
+  }).then( comment => {
+    //buscamos la pelicula en relacion al comentario
+    Movie.findByPk( movieId ).then( movie => {
+
+      const nunRate = parseInt(comment.rate);
+      const actualScore = movie.score;
+      res.json([nunRate, actualScore])
+      
+      //TODO: sacar el score y validar el rate
+
+      Movie.update({
+
+        score: 23 //pondremos el score total
+
+      }, {
+        where: {
+          id: movieId
+        }
+    
+      }).then( respuesta => {
+      // res.json(respuesta)
+
+      })
+
+
+    })
+
+  }).catch( err => {
+
+    res.json({
+      ok: false,
+      msg: err.errors[0].message
     });
 
   })
@@ -186,10 +307,13 @@ const deleteMovie = (req, res) => {
 }
 
 
+
 module.exports = {
-  getMovies,
-  createMovie,
-  getMovie,
-  updateMovie,
-  deleteMovie
+  getTheMovies,
+  createTheMovie,
+  getTheMovie,
+  updateTheMovie,
+  deleteTheMovie,
+  getMovieAndComments,
+  rateAndCommentMovie
 }
